requirements.txt
glfw==2.9.0
numpy==2.3.0
PyOpenGL==3.1.9
